Android Review and Update
The APP on mobile (Android) has some slight issues. The Front-end is sighty diffrente in terms of position. The Income Page isnt working on mobile and the profile page sends the "error" when trying to fetch the email.
And the navbar overlaps some elements since its always on the bottom of the phone but thats a minor inconvinience.
This fixes some of the front-end mainly on the Support and Settings Page.

(The Support Page is being fixed)